Move MaxWidthModifier to separate file for reusing
